requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
openpyxl>=3.0.0
matplotlib>=3.5.0
plotly>=5.0.0
numpy>=1.21.0
kaleido==0.2.1
openai>=1.0.0
python-dotenv>=1.0.0

